+++ src/main/java/com/example/youcodeRecruitment/Repository/HRRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface HRRepository extends JpaRepository<HR, Integer> {
+public interface HRRepository extends JpaRepository<HR, Long> {
     Optional<HR> findByEmail(String email);
 }
